+++ src/ol/events/ErrorEvent.js
@@ -1,5 +1,5 @@
 /**
- * @module ol-stac/events/ErrorEvent
+ * @module ol/events/ErrorEvent
  */
 import BaseEvent from 'ol/events/Event.js';
 import EventType from 'ol/events/EventType.js';