+++ src/ol/proj.js
@@ -1,5 +1,5 @@
 /**
- * @module ol-stac/proj
+ * @module ol/proj
  */
 
 import {