+++ src/ol/util.js
@@ -1,5 +1,5 @@
 /**
- * @module ol-stac/util
+ * @module ol/util
  */
 
 import VectorLayer from 'ol/layer/Vector.js';
@@ -8,6 +8,7 @@ import Fill from 'ol/style/Fill.js';
 import Stroke from 'ol/style/Stroke.js';
 import Style from 'ol/style/Style.js';
 import {VERSION} from 'ol/util.js';
+import {isObject} from 'stac-js/src/utils.js';
 
 /**
  * @typedef {import('ol/colorlike.js').ColorLike} ColorLike
@@ -161,7 +162,7 @@ export async function getStacObjectsForEvent(
 /**
  * Get the source info for the GeoTiff from the asset.
  * @param {import('stac-js').Asset} asset The asset to read the information from.
- * @param {Array<number>} selectedBands The (one-based) bands to show.
+ * @param {Array<number|string>} selectedBands The bands to show. One-based index of the band, or the name of the band.
  * @return {import('ol/source/GeoTIFF.js').SourceInfo} The source info for the GeoTiff asset
  */
 export function getGeoTiffSourceInfoFromAsset(asset, selectedBands) {
@@ -240,7 +241,22 @@ export function getGeoTiffSourceInfoFromAsset(asset, selectedBands) {
   }
 
   if (selectedBands.length > 0) {
-    sourceInfo.bands = selectedBands;
+    sourceInfo.bands = selectedBands
+      .map((band) => {
+        if (typeof band === 'number') {
+          return band;
+        }
+        const b = asset.findBand(band);
+        if (b) {
+          return b.getIndex() + 1;
+        }
+        // eslint-disable-next-line no-console
+        console.error(
+          `Band with name ${band} not found in asset ${asset.getKey()}`,
+        );
+        return null;
+      })
+      .filter((band) => band !== null);
   } else {
     const visualBands = asset.findVisualBands();
     if (visualBands) {
@@ -270,6 +286,51 @@ export function getBoundsStyle(originalStyle, layerGroup) {
     style.setFill(transparentFill);
   }
   return style;
+}
+
+/**
+ * Parse the GeoZarr source options from an asset.
+ *
+ * @param {Asset} asset The asset to read the information from.
+ * @param {Array<number|string>} selectedBands The bands to show. One-based index of the band, or the name of the band.
+ * @return {Object} The GeoZarr source options
+ * @api
+ */
+export function getGeoZarrSourceOptionsFromAsset(asset, selectedBands) {
+  const options = {
+    url: asset.getAbsoluteUrl(),
+  };
+
+  if (selectedBands.length > 0) {
+    options.bands = selectedBands
+      .map((band) => {
+        if (typeof band === 'string') {
+          return band;
+        }
+        const bands = asset.getBands();
+        const bandObj = bands[band - 1];
+        if (isObject(bandObj) && typeof bandObj.name === 'string') {
+          return bandObj.name;
+        }
+        // eslint-disable-next-line no-console
+        console.error(
+          `Band with index ${band} not found in asset ${asset.getKey()}`,
+        );
+        return null;
+      })
+      .filter(Boolean);
+  } else {
+    const bands = asset.findVisualBands();
+    if (bands) {
+      options.bands = [
+        bands.red.name,
+        bands.green.name,
+        bands.blue.name,
+      ].filter(Boolean);
+    }
+  }
+
+  return options;
 }
 
 /**