+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Support for WKT2 and PROJJSON (`proj:wkt2` and `proj:projjson`)
+- Removed GeoTIFF source and `loadProjection` helper in favor of ol
+- Render GeoTiffs better based on statistical values
+- Better per-band Visualization of GeoTiffs
+- Recommend usage of ol >= 10.9.0, require ol >= 9.0.0
+- Update stac-js
+- Fix examples
+
 ## [1.2.0] - 2026-03-13
 
 - Support for `classification:classes`